--- frontend/src/components/Hamburger/HamburgerMenuContent.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-const HamburgerMenuContent = (props) => {
-  return <div>Content</div>;
-};
-
-HamburgerMenuContent.propTypes = {};
-
-export default HamburgerMenuContent;
