+++ frontend/src/components/Hamburger/MenuButton/style.js
@@ -63,32 +63,3 @@ export const Container = styled.div`
     background-color: black;
   } */
 `;
-
-export const ModalBackGround = styled.div`
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.25);
-  position: fixed;
-  left: 0;
-  top: 0;
-`;
-
-export const ModalContainer = styled.div`
-  background-color: white;
-  position: absolute;
-  left: 100%;
-  padding-top: 50px;
-  /* 50%까지 이동. */
-  width: 50%;
-  height: 100%;
-  /* transform: ${(props) => (props.open ? 'translateX(-100%)' : '')}; */
-  transition: 0.4s;
-  &.modal-open {
-    transition: 0.4s;
-    transform: translateX(-100%);
-  }
-  &.modal-close {
-    transition: 0.4s;
-    transform: translateX(100%);
-  }
-`;
