+++ frontend/src/components/Hamburger/index.js
@@ -1,4 +1,4 @@
-import HamburgerMenu from './HamburgerMenu';
-import HamburgerModal from './HamburgerModal';
+import HamburgerMenu from './MenuButton';
+import HamburgerModal from './Modal';
 
 export { HamburgerMenu, HamburgerModal };
